Added Sitecore 'SecurityDisabler' around the item lookups.
The code could not have access to nodes for which the user isn't authenticated. In an extranet setting this means an item cannot be resolved for the current URL. Further down the pipeline a security resolver cannot check the attributes for that item and issue the correct 403, instead a 404, item not found will be served.